requirements.txt
# Python dependencies for ans_dev_sandbox_playbook
# This file contains all Python packages required for development and testing

# Core Ansible development tools
# Includes ansible-core, ansible-lint, molecule, and other essential tools
# Note: ansible-dev-tools installs pytest-ansible which conflicts with pytest-testinfra
# The activation script automatically uninstalls pytest-ansible after installation
ansible-dev-tools

# JMESPath for JSON query support in Ansible
jmespath

# Molecule plugins for testing
molecule-plugins

# Pytest infrastructure testing plugin for Molecule
pytest-testinfra

# YAML parsing library (required by DECRYPT_VAULTED_ITEMS.py)
PyYAML

# Syntax highlighting library (required by DECRYPT_VAULTED_ITEMS.py)
pygments
